Refactor: safety, performance, and clippy cleanup across 17 files
- Replace ~15 unwraps with if-let, unwrap_or, or expect
- Add with_capacity for ~12 known-size Vec allocations
- Use entry API for ~5 HashMap insert-or-update patterns
- Replace ~10 manual indexing loops with iterator chains
- Remove ~8 redundant clones, derive Copy for fieldless enums
- Apply ~160 clippy fixes: unnecessary casts, collapsible conditions,
  is_some_and, div_ceil, range contains, saturating_sub
- Accept &mut [u32] instead of &mut Vec<u32> in grid functions
- Remove dead code and unused variables

+++ src/types.rs
@@ -552,7 +552,7 @@ impl Blueprint {
     pub fn new(block_data: u32) -> Self {
         let bt = block_data & 0xFF;
         //                          (time, wood, clay, planks, rock, rope)
-        let (build_time, wood, clay, planks, rock, rope) = match bt as u32 {
+        let (build_time, wood, clay, planks, rock, rope) = match bt {
             //                     time  wood clay plnk rock rope
             // --- Floors ---
             BT_ROUGH_FLOOR => (0.8, 1, 0, 0, 0, 0),
@@ -666,7 +666,7 @@ impl Blueprint {
     }
 
     pub fn is_campfire(&self) -> bool {
-        let bt = (self.block_data & 0xFF) as u32;
+        let bt = self.block_data & 0xFF;
         (bt == BT_FIREPLACE || bt == BT_CAMPFIRE) && self.wall_edges == 0
     }
